+++ database/migrations/2016_10_06_191316_update_table_workouts_add_heart_rate_flds.php
@@ -27,8 +27,10 @@ class UpdateTableWorkoutsAddHeartRateFlds extends Migration
     public function down()
     {
         Schema::table('workouts', function($table){
-           $table->dropColumn('avg_hr');
-           $table->dropColumn('max_hr');
+            if ('sqlite' !== env('DB_CONNECTION')) {
+                $table->dropColumn('avg_hr');
+                $table->dropColumn('max_hr');
+            }
         });
     }
 }
